requirements.txt
# requirements.txt
python-dotenv
requests
openai
pdf2image